Refactor columns cuplication to preprocessing.schema
